some package renaming and restructuring

+++ scripts/cl-mpi-test.sh
@@ -8,6 +8,12 @@ then
     exit
 fi
 
+if [ ! `which cl-launch` ]
+then
+    echo "This script requires cl-launch (http://cliki.net/CL-Launch)"
+    exit.
+fi
+
 if [ $1 = "clean" ]
 then
     rm -f $BASEDIR/*-cl-mpi-image
@@ -30,7 +36,7 @@ for LISP in $LISP_IMPLEMENTATIONS; do
     echo "Building $LISP image $IMAGE..."
 
         cl-launch --lisp $LISP \
-                  -Q -s cl-mpi -s cl-mpi-testsuite -s mpi-benchmarks \
+                  -Q -s cl-mpi -s cl-mpi-testsuite -s cl-mpi-benchmarks \
                   -o $IMAGE -d $IMAGE -L $BASEDIR/cl-mpi-test.lisp -E main
     echo "...$LISP image complete"
     echo "========================================="
